Rename constructor argument compatibility fixture
