+++ tests/tests/src/constructor_payload_compatibility_test.mjs
@@ -2,24 +2,24 @@
 
 import * as Mocha from "mocha";
 import * as Test_utils from "./test_utils.mjs";
-import * as Constructor_explicit_arity from "./constructor_explicit_arity.mjs";
+import * as Constructor_arguments from "./constructor_arguments.mjs";
 
 Mocha.describe("constructor payload compatibility", () => {
   Mocha.test("accepts both unary tuple spellings", () => {
-    Test_utils.eq("File \"constructor_payload_compatibility_test.res\", line 7, characters 7-14", Constructor_explicit_arity.readUnary(Constructor_explicit_arity.unary), 3);
-    Test_utils.eq("File \"constructor_payload_compatibility_test.res\", line 8, characters 7-14", Constructor_explicit_arity.readUnary(Constructor_explicit_arity.unaryUnparenthesized), 3);
-    Test_utils.eq("File \"constructor_payload_compatibility_test.res\", line 9, characters 7-14", Constructor_explicit_arity.readUnaryUnparenthesized(Constructor_explicit_arity.unary), 3);
-    Test_utils.eq("File \"constructor_payload_compatibility_test.res\", line 10, characters 7-14", Constructor_explicit_arity.readUnaryUnparenthesized(Constructor_explicit_arity.unaryUnparenthesized), 3);
+    Test_utils.eq("File \"constructor_payload_compatibility_test.res\", line 7, characters 7-14", Constructor_arguments.readUnary(Constructor_arguments.unary), 3);
+    Test_utils.eq("File \"constructor_payload_compatibility_test.res\", line 8, characters 7-14", Constructor_arguments.readUnary(Constructor_arguments.unaryUnparenthesized), 3);
+    Test_utils.eq("File \"constructor_payload_compatibility_test.res\", line 9, characters 7-14", Constructor_arguments.readUnaryUnparenthesized(Constructor_arguments.unary), 3);
+    Test_utils.eq("File \"constructor_payload_compatibility_test.res\", line 10, characters 7-14", Constructor_arguments.readUnaryUnparenthesized(Constructor_arguments.unaryUnparenthesized), 3);
   });
   Mocha.test("accepts both binary constructor spellings", () => {
-    Test_utils.eq("File \"constructor_payload_compatibility_test.res\", line 13, characters 7-14", Constructor_explicit_arity.readBinary(Constructor_explicit_arity.binary), 3);
-    Test_utils.eq("File \"constructor_payload_compatibility_test.res\", line 14, characters 7-14", Constructor_explicit_arity.readBinary(Constructor_explicit_arity.binaryParenthesized), 3);
-    Test_utils.eq("File \"constructor_payload_compatibility_test.res\", line 15, characters 7-14", Constructor_explicit_arity.readBinaryParenthesized(Constructor_explicit_arity.binary), 3);
-    Test_utils.eq("File \"constructor_payload_compatibility_test.res\", line 16, characters 7-14", Constructor_explicit_arity.readBinaryParenthesized(Constructor_explicit_arity.binaryParenthesized), 3);
+    Test_utils.eq("File \"constructor_payload_compatibility_test.res\", line 13, characters 7-14", Constructor_arguments.readBinary(Constructor_arguments.binary), 3);
+    Test_utils.eq("File \"constructor_payload_compatibility_test.res\", line 14, characters 7-14", Constructor_arguments.readBinary(Constructor_arguments.binaryParenthesized), 3);
+    Test_utils.eq("File \"constructor_payload_compatibility_test.res\", line 15, characters 7-14", Constructor_arguments.readBinaryParenthesized(Constructor_arguments.binary), 3);
+    Test_utils.eq("File \"constructor_payload_compatibility_test.res\", line 16, characters 7-14", Constructor_arguments.readBinaryParenthesized(Constructor_arguments.binaryParenthesized), 3);
   });
   Mocha.test("accepts unparenthesized option tuple payloads", () => {
-    Test_utils.eq("File \"constructor_payload_compatibility_test.res\", line 19, characters 7-14", Constructor_explicit_arity.readOptionUnparenthesized(Constructor_explicit_arity.optionUnparenthesized), 3);
-    Test_utils.eq("File \"constructor_payload_compatibility_test.res\", line 20, characters 7-14", Constructor_explicit_arity.readOptionUnparenthesized([
+    Test_utils.eq("File \"constructor_payload_compatibility_test.res\", line 19, characters 7-14", Constructor_arguments.readOptionUnparenthesized(Constructor_arguments.optionUnparenthesized), 3);
+    Test_utils.eq("File \"constructor_payload_compatibility_test.res\", line 20, characters 7-14", Constructor_arguments.readOptionUnparenthesized([
       1,
       2
     ]), 3);
